Add extra fields for presidential candidates to the edit page
For Burkina Faso, it's been requested that we add a "CV" and "Program"
fields for candidates who are standing in the presidential election.

This doesn't seem like a particularly good approach (e.g. the
optional 'cv' and 'program' fields being directly referenced in
'popit_data') but it will probably do for the moment.