Only show top N candidates on party lists
Adds a max_count param to group_people_by_party
Limits the number of members returned per party to max_count, defaults
to None, uses the default_party_list_members_to_show setting for
max_count.
Party object passed to template includes the max_count and, if the list
has been truncated the total count in the result. Uses the total count
setting to show/hide the link to the full party list in the template.
Also now says how many people are in the full list.

Sorts the members of a list by list position.
If we're using party lists then we want to see the candidates in the
order they appear in the party list rather than alphabetical, and this
also means that if we are truncating the list we are showing the top N
candidates on the list, and hence most likely to be elected  and not
the first N alphabetically.

We make ths list of people a list of tuples of ( list_postion, person )
and then either sort on the first item or the last_name property of the
second.

When we construct the result we throw away the first item in the tuple
and return just a list of person objects because we don't need the
position