sort by party list position if using party lists
If we're using party lists then we want to see the candidates in the
order they appear in the party list rather than alphabetical, and this
also means that if we are truncating the list we are showing the top N
candidates on the list, and hence most likely to be elected  and not
the first N alphabetically.

We make ths list of people a list of tuples of ( list_postion, person )
and then either sort on the first item or the last_name property of the
second.

When we construct the result we throw away the first item in the tuple
and return just a list of person objects because we don't need the
position